requirements.txt
appdirs==1.4.4
astroid==2.11.6
autopep8==1.6.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cssselect==1.1.0
dill==0.3.5.1
dnspython==2.3.0
exceptiongroup==1.1.1
fake-useragent==0.1.11
Flask==2.1.2
idna==3.3
importlib-metadata==4.12.0
iniconfig==2.0.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
libretranslatepy==2.1.1
lxml==4.9.1
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==23.1
parse==1.19.0
platformdirs==2.5.2
pluggy==1.0.0
pycodestyle==2.8.0
pyee==8.2.2
pylint==2.14.4
pymongo==4.3.3
pyppeteer==1.0.2
pyquery==1.4.3
pytest==7.3.1
python-dotenv==0.20.0
requests==2.31.0
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
tqdm==4.64.0
translate==3.6.1
typing_extensions==4.3.0
Unidecode==1.3.4
urllib3==1.26.9
w3lib==1.22.0
websockets==10.3
Werkzeug==2.3.6
wrapt==1.14.1
zipp==3.8.0
